combine test-driver and frame-utils to utils.js
